rename 'd_diffusion_sa' term to 'd_sd_diffusion', remove unused functions
